Move FilterButton component
